+++ src/main/scala/org/quantlib/time/daycounts/One.scala
@@ -5,7 +5,7 @@ import org.quantlib.time.implicits.DateOps._
 
 final case class One[D: DateOps]() extends DayCountBasis[D] {
 
-  override def dayCount(date1: D, date2: D): Long = if (date1 > date2) -1 else 1
+  override def dayCount(date1: D, date2: D): Int = if (date1 > date2) -1 else 1
 
   override def yearFraction(date1: D, date2: D,
                    refDate1: Option[D] = None,
